Fix Part of #4236 : Add tests for AsyncResultSubject (#5670)
## Explanation
Fixes part of
[#4236](https://github.com/oppia/oppia-android/issues/4236): Adds tests
to ensure correct handling of non-iterable async results.

This PR adds tests for the `AsyncResultSubject` class. It verifies that
the `AsyncResultSubject` handles non-iterable results correctly and
ensures that the expected failures occur when the result is not
iterable. The test specifically checks that the appropriate behavior is
triggered when a non-iterable result is provided.

Updated Tests : 


![Image](https://github.com/user-attachments/assets/d09dcc6a-b533-4073-a6ca-0c9328dcf3d8)

## Essential Checklist
<!-- Please tick the relevant boxes by putting an "x" in them. -->
- [x] The PR title and explanation each start with "Fix #bugnum: " (If
this PR fixes part of an issue, prefix the title with "Fix part of
#bugnum: ...".)
- [ ] Any changes to
[scripts/assets](https://github.com/oppia/oppia-android/tree/develop/scripts/assets)
files have their rationale included in the PR explanation.
- [x] The PR follows the [style
guide](https://github.com/oppia/oppia-android/wiki/Coding-style-guide).
- [x] The PR does not contain any unnecessary code changes from Android
Studio
([reference](https://github.com/oppia/oppia-android/wiki/Guidance-on-submitting-a-PR#undo-unnecessary-changes)).
- [x] The PR is made from a branch that's **not** called "develop" and
is up-to-date with "develop".
- [x] The PR is **assigned** to the appropriate reviewers
([reference](https://github.com/oppia/oppia-android/wiki/Guidance-on-submitting-a-PR#clarification-regarding-assignees-and-reviewers-section)).